+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.16.0] - 2026-06-28
+## [1.17.0] - 2026-06-28
+
+### Added
+
+- Date range filter and URL path search on analytics page
+- Pagination on Top Pages (25 per page) and Recent Visitors (50 per page)
+- Quick date buttons: Today, This Week, This Month, Clear
+
+## [1.16.0] - 2026-06-28 ([65d524f](https://github.com/anomalyco/cleantouch/commit/65d524f))
 
 ### Changed